Improved validation error details
The `RSchema::ErrorDetails` class has been improved to be both more
human friendly, and machine readable. This will break backwards
compatibility with code that expects `RSchema::ErrorDetails#details` to
exist, but should be fine otherwise.

Also, `RSchema.validation_errors` was renamed to 
`RSchema.validation_error` because it returns a single object, not a
collection.