Improved error objects for VariableHash schemas
Previously, VariableHash schemas would return an error object whenever
either of the subschemas failed. Users would probably expect a hash
instead, since this is how all the other schema types work.

Now, if either of the subschemas returns an invalid result, the error
object will always be a hash with two elements: `:keys` and `:values`,
which are hashes of error objects themselves.

 - For invalid values, check `result.error[:values][:name]`.
 - For invalid keys, check `result.error[:keys][:name]`.

The error hash doesn't exactly mirror the structure of the input value,
but there is no other way to represents _both_ key and value errors.