Allow HTTPCoercer to be double wrapped
HTTPCoercer creates a new schema by wrapping an existing schema.
This new wrapped schema can be wrapped _again_ by HTTPCoercer and
it will work fine.

There is no real reason to do this with the same wrapper, but it will
allow different wrappers to work with each other in the future instead
of just crashing.