Standardise RSchema::Error#vars
Error vars now must be a hash. The elements of the hash should be
treated as extra variables that describe why the error occurred.

The `Sum` and `Set` schemas no longer use error vars to hold their
subschema errors. Instead, the subschema errors are now part of the
resulting error hash.

There are now no schemas that store subschema errors in the error vars.