Improvements to RACK_PARAMS coercion wrapper
Defaults arrays to empty arrays. Due to the way Rack params are made
from the form-encoded params, empty arrays do not turn up in the
resulting param hash (their key is missing). Now they will turn up in
the param hash, as an empty array.

Defaults empty strings to nil, for `maybe(...)` schemas. If you have a
HTML <select> field with a selected option like "None", the field is
still submitted with an empty string as the value. Now, empty strings
will be coerced to nil where `maybe` schemes are present.